Move `StringInreplaceExtension` into separate file.
